Refactor mod sorting logic into separate module
- Extract sorting-related functions and classes from mods_panel.py into new app/sort/mod_sorting.py
- Update imports in main_content_panel.py and mods_panel.py to use the new module
- Improve code organization and maintainability
